+++ apps/rest-api-server/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @blobscan/rest-api-server
 
+## 0.3.5
+
+### Patch Changes
+
+- Updated dependencies [[`37539a6`](https://github.com/Blobscan/blobscan/commit/37539a6881e1cffe2cd3e7ef6f1686e6d6f39bd7)]:
+  - @blobscan/api@0.6.0
+
 ## 0.3.4
 
 ### Patch Changes