+++ packages/blob-storage-manager/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @blobscan/blob-storage-manager
 
+## 0.1.0
+
+### Minor Changes
+
+- [#355](https://github.com/Blobscan/blobscan/pull/355) [`37539a6`](https://github.com/Blobscan/blobscan/commit/37539a6881e1cffe2cd3e7ef6f1686e6d6f39bd7) Thanks [@0xGabi](https://github.com/0xGabi)! - Added new procedure to fetch swarm state
+
+## 0.0.8
+
+### Patch Changes
+
+- [#353](https://github.com/Blobscan/blobscan/pull/353) [`fc6298f`](https://github.com/Blobscan/blobscan/commit/fc6298fcbdc17b89bd0289ddd1f8d252870cd402) Thanks [@PJColombo](https://github.com/PJColombo)! - Fixed swarm storage blob data decoding
+
 ## 0.0.7
 
 ### Patch Changes